Move schema info logic to driver implementation
