+++ server/drivers/mssql.js
@@ -1,24 +1,23 @@
 const mssql = require('mssql')
 const QueryResult = require('../models/QueryResult')
+const { formatSchemaQueryResults } = require('./utils')
 
-function getSchemaSql() {
-  return `
-    SELECT 
-      t.table_schema, 
-      t.table_name, 
-      c.column_name, 
-      c.data_type
-    FROM 
-      INFORMATION_SCHEMA.TABLES t 
-      JOIN INFORMATION_SCHEMA.COLUMNS c ON t.table_schema = c.table_schema AND t.table_name = c.table_name 
-    WHERE 
-      t.table_schema NOT IN ('information_schema') 
-    ORDER BY 
-      t.table_schema, 
-      t.table_name, 
-      c.ordinal_position
-  `
-}
+const SCHEMA_SQL = `
+  SELECT 
+    t.table_schema, 
+    t.table_name, 
+    c.column_name, 
+    c.data_type
+  FROM 
+    INFORMATION_SCHEMA.TABLES t 
+    JOIN INFORMATION_SCHEMA.COLUMNS c ON t.table_schema = c.table_schema AND t.table_name = c.table_name 
+  WHERE 
+    t.table_schema NOT IN ('information_schema') 
+  ORDER BY 
+    t.table_schema, 
+    t.table_name, 
+    c.ordinal_position
+`
 
 function runQuery(query, connection) {
   const queryResult = new QueryResult()
@@ -114,8 +113,18 @@ function testConnection(connection) {
   return runQuery(query, connection)
 }
 
+/**
+ * Get schema for connection
+ * @param {*} connection
+ */
+function getSchema(connection) {
+  return runQuery(SCHEMA_SQL, connection).then(queryResult =>
+    formatSchemaQueryResults(queryResult)
+  )
+}
+
 module.exports = {
-  getSchemaSql,
+  getSchema,
   runQuery,
   testConnection
 }
